requirements.txt
requests
telebot
systemd-python
python-dotenv